refactor(test-package): 중복되는 패키지를 삭제합니다.
